Update Help
As per this email request "Dear Contributor, please click just before the first speech mark on line 9 and press the Backspace key once so that it is parallel with line 8. As well as this, please add your policy ID to your twitter so that I can verify." I didn't see that it was not parallel. I used json formatter and resubmitting. Thanks for your patience.